Module calloop::error

Error types used and generated by Calloop.

This module contains error types for Calloop’s operations. They are designed to make it easy to deal with errors arising from Calloop’s internal I/O and other operations.

There are two top-level error types:

Error: used by callback functions, internal operations, and some event loop API calls

InsertError: used primarily by the insert_source() method when an event source cannot be added to the loop and needs to be given back to the caller

Structs

InsertError

An error generated when trying to insert an event source

Enums

Error

The primary error type used by Calloop covering internal errors and I/O errors that arise during loop operations such as source registration or event dispatching.

Type Definitions

Result

Result alias using Calloop’s error type.