Struct calloop::generic::Generic

pub struct Generic<F: AsRawFd> {
    pub file: F,
    pub interest: Interest,
    pub mode: Mode,
}

A generic event source wrapping a FD-backed type

Fields
file: F

The wrapped FD-backed type

interest: Interest

The programmed interest

mode: Mode

The programmed mode

Implementations

impl<F: AsRawFd + 'static> Generic<F>

pub fn new(file: F, interest: Interest, mode: Mode) -> Generic<F>

Wrap a FD-backed type into a Generic event source

pub fn unwrap(self) -> F

Unwrap the Generic source to retrieve the underlying type

impl Generic<Fd>

pub fn from_fd(fd: RawFd, interest: Interest, mode: Mode) -> Generic<Fd>

Wrap a raw file descriptor into a Generic event source