Struct calloop::io::Async

pub struct Async<'l, F: AsRawFd> { /* fields omitted */ }

Adapter for async IO manipulations

This type wraps an IO object, providing methods to create futures waiting for its readiness.

If the futures-io cargo feature is enabled, it also implements AsyncRead and/or AsyncWrite if the underlying type implements Read and/or Write.

Note that this adapter and the futures procuded from it and not threadsafe.

Implementations

impl<'l, F: AsRawFd> Async<'l, F>

pub fn get_mut(&mut self) -> &mut F

Mutably access the underlying IO object

pub fn readable<'s>(&'s mut self) -> Readable<'s, 'l, F>

A future that resolves once the object becomes ready for reading

pub fn writable<'s>(&'s mut self) -> Writable<'s, 'l, F>

A future that resolves once the object becomes ready for writing

pub fn into_inner(self) -> F

Remove the async adapter and retrieve the underlying object