calloop

Module loop_logic

Structs

EventIterator

The EventIterator is an Iterator over the events relevant to a particular source This type is used in the EventSource::before_handle_events methods for two main reasons:

EventLoop

An event loop

LoopHandle

An handle to an event loop

LoopInner 🔒

LoopSignal

A signal that can be shared between thread to stop or wakeup a running event loop

RegistrationToken

A token representing a registration in the EventLoop.

Signals 🔒

Signals related to the event loop.

Type Aliases

IdleCallback 🔒