calloop::ping

Type Alias PingSource

pub type PingSource = PingSource;

The Ping handle

This handle can be cloned and sent accross threads. It can be used to send pings to the PingSource.

Aliased Type

pub struct PingSource {
    event: Generic<ArcAsFd>,
}

Fields
event: Generic<ArcAsFd>