Struct calloop::signals::Signals

pub struct Signals { /* fields omitted */ }

An event source for receiving Unix signals

Implementations

impl Signals

pub fn new(signals: &[Signal]) -> Result<Signals>

Create a new signal event source listening on the specified list of signals

pub fn add_signals(&mut self, signals: &[Signal]) -> Result<()>

Add a list of signals to the signals source

If this function returns an error, the signal mask of the thread may have still been changed.

pub fn remove_signals(&mut self, signals: &[Signal]) -> Result<()>

Remove a list of signals to the signals source

If this function returns an error, the signal mask of the thread may have still been changed.

pub fn set_signals(&mut self, signals: &[Signal]) -> Result<()>

Replace the list of signals of the source

If this function returns an error, the signal mask of the thread may have still been changed.