Struct calloop::timer::Timer

pub struct Timer<T> { /* fields omitted */ }

A Timer event source

It generates events of type (T, TimerHandle<T>), providing you an handle inside the event callback, allowing you to set new timeouts as a response to a timeout being reached (for reccuring ticks for example).

Methods

impl<T> Timer<T>

pub fn new() -> Timer<T>

Create a new timer with default parameters

Default time resolution is 100ms

pub fn with_resolution(resolution: Duration) -> Timer<T>

Create a new timer with a specific time resolution

pub fn handle(&self) -> TimerHandle<T>

Get an handle for this timer