Trait calloop::EventDispatcher

pub trait EventDispatcher<Data> {
    fn ready(&mut self, ready: Ready, data: &mut Data);
}

An event dispatcher

It is the junction between user callbacks and and an event source, receiving mio readinesses, converting them into appropriate events and calling their inner user callback.

Required methods

fn ready(&mut self, ready: Ready, data: &mut Data)

The source has a readiness event

Loading content...