Module calloop::channel

An MPSC channel whose receiving end is an event source

Create a channel using Channel::<T>::new(), which returns a Sender<T> that can be cloned and sent accross threads if T: Send, and a Channel<T> that can be inserted into an EventLoop. It will generate one event per message.

This implementation is based on mio_more::channel.

Structs

Channel
The receiving end of the channel

Sender
The sending half of a channel.

SyncSender
The sending half of a synchronous channel.

Enums

Event
The events generated by the channel event source

SendError
An error returned from the Sender::send or SyncSender::send function.

TrySendError
An error returned from the SyncSender::try_send function.

Functions

channel
Create a new asynchronous channel

sync_channel
Create a new synchronous, bounded channel