calloop::ping

Function make_ping

pub fn make_ping() -> Result<(Ping, PingSource)>

Create a new ping event source

you are given a Ping instance, which can be cloned and used to ping the event loop, and a PingSource, which you can insert in your event loop to receive the pings.