Type Alias calloop::ping::PingSource

pub type PingSource = PingSource;

The Ping handle

This handle can be cloned and sent accross threads. It can be used to send pings to the PingSource.

Aliased Type

struct PingSource {
    event: Generic<i32, Error>,
    _fd: Arc<CloseOnDrop, Global>,
}

Fields
event: Generic<i32, Error>_fd: Arc<CloseOnDrop, Global>