Module calloop::signals

Event source for tracking Unix signals

Only available on Linux.

This allows you to track and receive Unix signals through the event loop rather than by registering signal handlers. It uses signalfd under the hood.

The source will take care of masking and unmasking signals for the thread it runs on, but you are responsible for masking them on other threads if you run them. The simplest way to ensure that is to setup the signal event source before spawning any thread, as they’ll inherit their parent signal mask.

Structs

Event

An event generated by the signal event source

SignalError

An error arising from processing events for a process signal.

Signals

An event source for receiving Unix signals

siginfo

Enums

Signal

Types of operating system signals