Module calloop::sources

Modules

channel

An MPSC channel whose receiving end is an event source

generic

A generic event source wrapping an IO objects or file descriptor

ping

Ping to the event loop

signals

Event source for tracking Unix signals

timer

Timer event source

transient

Wrapper for a transient Calloop event source.

Structs

Dispatcher

An event source with its callback.

DispatcherInner 🔒

Idle

An idle callback that was inserted in this loop

Enums

PostAction

Possible actions that can be requested to the event loop by an event source once its events have been processed.

Traits

CancellableIdle 🔒

ErasedDispatcher 🔒

EventDispatcher 🔒

EventSource

Trait representing an event source

IdleDispatcher 🔒