Module calloop::sources

Modules

channel

An MPSC channel whose receiving end is an event source

generic

A generic event source wrapping a file descriptor

ping

Ping to the event loop

signals

Event source for tracking Unix signals

timer

Timer-based event sources

Structs

Dispatcher

Idle

An idle callback that was inserted in this loop

Source

A token representing an event source inserted in the event loop

Traits

CancellableIdle

EventDispatcher

EventSource

Trait representing an event source

IdleDispatcher