Struct calloop::EventLoop

pub struct EventLoop<Data> { /* fields omitted */ }

An event loop

This loop can host several event sources, that can be dynamically added or removed.

Methods

impl<Data: 'static> EventLoop<Data>

pub fn new() -> Result<EventLoop<Data>>

Create a new event loop

It is backed by an mio provided machinnery, and will fail if the mio initialization fails.

pub fn handle(&self) -> LoopHandle<Data>

Retrieve a loop handle

pub fn dispatch(
&mut self,
timeout: Option<Duration>,
data: &mut Data
) -> Result<()>

Dispatch pending events to their callbacks

Some source have events available, their callbacks will be immediatly called. Otherwise this will wait until an event is receive or the provided timeout is reached. If timeout is None, it will wait without a duration limit.

Once pending events have been processed or the timeout is reached, all pending idle callbacks will be fired before this method returns.

pub fn get_signal(&self) -> LoopSignal

Get a signal to stop this event loop from running

To be used in conjunction with the run() method.

pub fn run<F>(
&mut self,
timeout: Option<Duration>,
data: &mut Data,
cb: F
) -> Result<()> where
F: FnMut(&mut Data),

Run this event loop

This will repeatedly try to dispatch events (see the dispatch() method) on this event loop, waiting at most timeout every time.

Between each dispatch wait, your provided callback will be called.

You can use the get_signal() method to retrieve a way to stop or wakeup the event loop from anywhere.