calloop

Struct EventLoop

pub struct EventLoop<'l, Data> { /* private fields */ }

An event loop

This loop can host several event sources, that can be dynamically added or removed.

Implementations

impl<'l, Data> EventLoop<'l, Data>

pub fn try_new() -> Result<Self>

Create a new event loop

Fails if the initialization of the polling system failed.

pub fn handle(&self) -> LoopHandle<'l, Data>

Retrieve a loop handle

pub fn dispatch<D: Into<Option<Duration>>>( &mut self, timeout: D, data: &mut Data, ) -> Result<()>

Dispatch pending events to their callbacks

If some sources have events available, their callbacks will be immediately called. Otherwise, this will wait until an event is received or the provided timeout is reached. If timeout is None, it will wait without a duration limit.

Once pending events have been processed or the timeout is reached, all pending idle callbacks will be fired before this method returns.

pub fn get_signal(&self) -> LoopSignal

Get a signal to stop this event loop from running

To be used in conjunction with the run() method.

pub fn run<F, D: Into<Option<Duration>>>( &mut self, timeout: D, data: &mut Data, cb: F, ) -> Result<()>
where F: FnMut(&mut Data),

Run this event loop

This will repeatedly try to dispatch events (see the dispatch() method) on this event loop, waiting at most timeout every time.

Between each dispatch wait, your provided callback will be called.

You can use the get_signal() method to retrieve a way to stop or wakeup the event loop from anywhere.