Struct calloop::timer::Timer

pub struct Timer { /* private fields */ }

A timer event source

When registered to the event loop, it will trigger an event once its deadline is reached. If the deadline is in the past relative to the moment of its insertion in the event loop, the TImer will trigger an event as soon as the event loop is dispatched.

Implementations

impl Timer

pub fn immediate() -> Timer

Create a timer that will fire immediately when inserted in the event loop

pub fn from_duration(duration: Duration) -> Timer

Create a timer that will fire after a given duration from now

pub fn from_deadline(deadline: Instant) -> Timer

Create a timer that will fire at a given instant

pub fn set_deadline(&mut self, deadline: Instant)

Changes the deadline of this timer to an Instant

If the Timer is currently registered in the event loop, it needs to be re-registered for this change to take effect.

pub fn set_duration(&mut self, duration: Duration)

Changes the deadline of this timer to a Duration from now

If the Timer is currently registered in the event loop, it needs to be re-registered for this change to take effect.

pub fn current_deadline(&self) -> Instant

Get the current deadline of this Timer