Module calloop::transient

Wrapper for a transient Calloop event source.

If you have high level event source that you expect to remain in the event loop indefinitely, and another event source nested inside that one that you expect to require removal or disabling from time to time, this module can handle it for you.

Enums

TransientSource

A TransientSource wraps a Calloop event source and manages its registration. A user of this type only needs to perform the usual Calloop calls (process_events() and *register()) and the return value of process_events().